+++ packages/select/src/components/checkmark/Component.tsx
@@ -1,9 +1,7 @@
 import React, { useCallback } from 'react';
 import cn from 'classnames';
 
-import { Badge } from '@alfalab/core-components-badge';
 import { Checkbox, type CheckboxProps } from '@alfalab/core-components-checkbox';
-import { CheckmarkCircleMIcon } from '@alfalab/icons-glyph/CheckmarkCircleMIcon';
 import CheckmarkMIcon from '@alfalab/icons-glyph/CheckmarkMIcon';
 
 import { type CheckmarkProps } from '../../typings';
@@ -16,6 +14,7 @@ export const Checkmark = ({
     className,
     multiple,
     position = 'before',
+    content,
 }: CheckmarkProps) => {
     const single = !multiple;
 
@@ -32,17 +31,7 @@ export const Checkmark = ({
 
     return single ? (
         <div className={checkmarkClassNames}>
-            {position === 'before' ? (
-                <Badge
-                    className={styles.after}
-                    view='icon'
-                    size='m'
-                    iconColor='positive'
-                    content={<CheckmarkCircleMIcon className={styles.colorIcon} />}
-                />
-            ) : (
-                <CheckmarkMIcon className={cn(styles.displayIcon)} />
-            )}
+            <CheckmarkMIcon className={cn(styles.displayIcon)} />
         </div>
     ) : (
         <Checkbox
@@ -50,6 +39,7 @@ export const Checkmark = ({
             disabled={disabled}
             className={checkmarkClassNames}
             size={24}
+            label={content}
             onClick={handleCheckboxClick}
         />
     );