+++ scripts/demo-params/resolve-variant-directory/index.cjs
@@ -1,0 +1,8 @@
+// @ts-check
+
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-expect-error
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const { default: resolveVariantDirectory } = require('./index.js');
+
+module.exports = resolveVariantDirectory;