+++ packages/attach/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @alfalab/core-components-attach
 
+## 8.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-button@13.1.0
+
+## 8.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-button@13.0.2
+
 ## 8.0.1
 
 ### Patch Changes