+++ packages/calendar-range/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @alfalab/core-components-calendar-range
 
+## 9.0.5
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@9.1.0
+    - @alfalab/core-components-date-input@6.0.4
+    - @alfalab/core-components-calendar-input@12.0.5
+
+## 9.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-date-input@6.0.3
+    - @alfalab/core-components-calendar-input@12.0.4
+
+## 9.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@2.1.0
+    - @alfalab/core-components-date-input@6.0.2
+    - @alfalab/core-components-calendar@9.0.3
+    - @alfalab/core-components-calendar-input@12.0.3
+
+## 9.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@9.0.2
+    - @alfalab/core-components-calendar-input@12.0.2
+
 ## 9.0.1
 
 ### Patch Changes