+++ packages/circular-progress-bar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @alfalab/core-components-circular-progress-bar
 
+## 5.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@2.1.0
+    - @alfalab/core-components-typography@6.0.2
+
 ## 5.0.1
 
 ### Patch Changes