+++ packages/custom-button/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @alfalab/core-components-custom-button
 
+## 5.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-button@13.1.0
+
+## 5.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-button@13.0.2
+
 ## 5.0.1
 
 ### Patch Changes