+++ packages/custom-picker-button/CHANGELOG.md
@@ -1,5 +1,74 @@
 # @alfalab/core-components-custom-picker-button
 
+## 4.1.0
+
+### Minor Changes
+
+<sup><time>27.02.2026</time></sup>
+
+#### [#2071](https://github.com/core-ds/core-components/pull/2071)
+
+##### Select
+
+- Добавлен проп `checkmarkPosition` для управления позицией чекмарка относительно контента ('before' | 'after')
+- Компонент `BaseCheckmark` теперь поддерживает проп `content` для передачи контента опции
+- Оптимизирована логика рендеринга чекбоксов и иконок в компонентах `OptionBase` и `BaseOption`
+
+##### SelectWithTags
+
+- Добавлена поддержка динамического позиционирования чекмарков в зависимости от `view` (desktop/mobile)
+- Для desktop по умолчанию используется `checkmarkPosition='before'`, для mobile - `checkmarkPosition='after'`
+- Возможность переопределить позицию через проп `optionProps.checkmarkPosition`
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.1.0
+    - @alfalab/core-components-custom-button@5.0.3
+    - @alfalab/core-components-picker-button@13.0.7
+
+## 4.0.6
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.0.6
+    - @alfalab/core-components-picker-button@13.0.6
+
+## 4.0.5
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.0.5
+    - @alfalab/core-components-mq@6.0.2
+    - @alfalab/core-components-picker-button@13.0.5
+    - @alfalab/core-components-custom-button@5.0.2
+
+## 4.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.0.4
+    - @alfalab/core-components-picker-button@13.0.4
+
+## 4.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.0.3
+    - @alfalab/core-components-picker-button@13.0.3
+
+## 4.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.0.2
+    - @alfalab/core-components-picker-button@13.0.2
+
 ## 4.0.1
 
 ### Patch Changes