+++ packages/date-time-input/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @alfalab/core-components-date-time-input
 
+## 6.0.5
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.2
+    - @alfalab/core-components-calendar@9.1.0
+    - @alfalab/core-components-icon-button@8.0.3
+
+## 6.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.1
+
+## 6.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.0
+    - @alfalab/core-components-calendar@9.0.3
+    - @alfalab/core-components-mq@6.0.2
+    - @alfalab/core-components-popover@8.0.2
+    - @alfalab/core-components-icon-button@8.0.2
+
+## 6.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@9.0.2
+
 ## 6.0.1
 
 ### Patch Changes