+++ packages/gallery/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @alfalab/core-components-gallery
 
+## 7.0.6
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-button@13.1.0
+    - @alfalab/core-components-icon-button@8.0.3
+    - @alfalab/core-components-tooltip@9.0.6
+
+## 7.0.5
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-tooltip@9.0.5
+    - @alfalab/core-components-base-modal@7.1.1
+    - @alfalab/core-components-button@13.0.2
+    - @alfalab/core-components-mq@6.0.2
+    - @alfalab/core-components-spinner@6.0.2
+    - @alfalab/core-components-typography@6.0.2
+    - @alfalab/core-components-icon-button@8.0.2
+
+## 7.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-base-modal@7.1.0
+    - @alfalab/core-components-tooltip@9.0.4
+
+## 7.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-tooltip@9.0.3
+
+## 7.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-tooltip@9.0.2
+
 ## 7.0.1
 
 ### Patch Changes