+++ packages/input-autocomplete/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @alfalab/core-components-input-autocomplete
 
+## 14.0.7
+
+### Patch Changes
+
+<sup><time>27.02.2026</time></sup>
+
+#### [#2053](https://github.com/core-ds/core-components/pull/2053)
+
+##### Input, AmountInput, InputAutocomplete, NumberInput, Select, FormControl, SelectWithTags
+
+- Небольшие улучшения кодовой базы компонентов
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.2
+    - @alfalab/core-components-select@19.1.0
+    - @alfalab/core-components-form-control@14.0.3
+
+## 14.0.6
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.1
+    - @alfalab/core-components-select@19.0.6
+
+## 14.0.5
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.0
+    - @alfalab/core-components-shared@2.1.0
+    - @alfalab/core-components-select@19.0.5
+    - @alfalab/core-components-form-control@14.0.2
+    - @alfalab/core-components-mq@6.0.2
+    - @alfalab/core-components-popover@8.0.2
+
+## 14.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.0.4
+
+## 14.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.0.3
+
+## 14.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.0.2
+
 ## 14.0.1
 
 ### Patch Changes