+++ packages/input/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @alfalab/core-components-input
 
+## 17.1.2
+
+### Patch Changes
+
+<sup><time>27.02.2026</time></sup>
+
+#### [#2084](https://github.com/core-ds/core-components/pull/2084)
+
+##### Input
+
+- Исправлен цвет `placeholder` в `readOnly` состоянии
+
+<sup><time>27.02.2026</time></sup>
+
+#### [#2053](https://github.com/core-ds/core-components/pull/2053)
+
+##### Input, AmountInput, InputAutocomplete, NumberInput, Select, FormControl, SelectWithTags
+
+- Небольшие улучшения кодовой базы компонентов
+
+- Обновлены зависимости
+    - @alfalab/core-components-button@13.1.0
+    - @alfalab/core-components-form-control@14.0.3
+
+## 17.1.1
+
+### Patch Changes
+
+<sup><time>24.02.2026</time></sup>
+
+#### [#2075](https://github.com/core-ds/core-components/pull/2075)
+
+##### Input
+
+- Повышена специфичность класса `.focusVisible`, проставляемого при фокусе через клавиатуру
+
+## 17.1.0
+
+### Minor Changes
+
+<sup><time>20.02.2026</time></sup>
+
+#### [#2076](https://github.com/core-ds/core-components/pull/2076)
+
+##### Input
+
+- Добавлен проп `fontWeight` для управления жирностью шрифта
+- Проп `bold` помечен как `deprecated`. Используйте проп `fontWeight`
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@2.1.0
+    - @alfalab/core-components-button@13.0.2
+    - @alfalab/core-components-form-control@14.0.2
+    - @alfalab/core-components-mq@6.0.2
+
 ## 17.0.1
 
 ### Patch Changes