+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @alfalab/core-components-modal
 
+## 11.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-navigation-bar-private@2.0.4
+
+## 11.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@2.1.0
+    - @alfalab/core-components-base-modal@7.1.1
+    - @alfalab/core-components-mq@6.0.2
+    - @alfalab/core-components-navigation-bar-private@2.0.3
+
+## 11.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-base-modal@7.1.0
+    - @alfalab/core-components-navigation-bar-private@2.0.2
+
 ## 11.0.1
 
 ### Patch Changes