+++ packages/navigation-bar/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alfalab/core-components-navigation-bar
 
+## 3.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-button@13.1.0
+
+## 3.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@2.1.0
+    - @alfalab/core-components-button@13.0.2
+    - @alfalab/core-components-typography@6.0.2
+
 ## 3.0.1
 
 ### Patch Changes