+++ packages/notification/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @alfalab/core-components-notification
 
+## 9.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-toast-plate@9.0.3
+
+## 9.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@2.1.0
+    - @alfalab/core-components-toast-plate@9.0.2
+
 ## 9.0.1
 
 ### Patch Changes