+++ packages/number-input/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @alfalab/core-components-number-input
 
+## 4.0.4
+
+### Patch Changes
+
+<sup><time>27.02.2026</time></sup>
+
+#### [#2053](https://github.com/core-ds/core-components/pull/2053)
+
+##### Input, AmountInput, InputAutocomplete, NumberInput, Select, FormControl, SelectWithTags
+
+- Небольшие улучшения кодовой базы компонентов
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.2
+    - @alfalab/core-components-icon-button@8.0.3
+
+## 4.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.1
+
+## 4.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.0
+    - @alfalab/core-components-shared@2.1.0
+    - @alfalab/core-components-mq@6.0.2
+    - @alfalab/core-components-icon-button@8.0.2
+
 ## 4.0.1
 
 ### Patch Changes