+++ packages/phone-input/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @alfalab/core-components-phone-input
 
+## 9.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-masked-input@8.0.4
+
+## 9.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-masked-input@8.0.3
+
+## 9.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-masked-input@8.0.2
+
 ## 9.0.1
 
 ### Patch Changes