+++ packages/slider-input/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @alfalab/core-components-slider-input
 
+## 11.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.2
+
+## 11.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.1
+
+## 11.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.0
+
 ## 11.0.1
 
 ### Patch Changes