+++ packages/steps/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alfalab/core-components-steps
 
+## 3.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@2.1.0
+
 ## 3.0.1
 
 ### Patch Changes