+++ packages/textarea/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alfalab/core-components-textarea
 
+## 10.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-form-control@14.0.3
+
+## 10.0.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@2.1.0
+    - @alfalab/core-components-form-control@14.0.2
+    - @alfalab/core-components-mq@6.0.2
+
 ## 10.0.1
 
 ### Patch Changes